[REFACTO] Hour by hour ram/Cpu From Service to Server_base_class + Tests
