+++ efootprint/core/service.py
@@ -14,11 +14,6 @@ class Service(ModelingObject):
                  base_cpu_consumption: SourceValue = None):
         super().__init__(name)
         self.storage_needed = None
-        self.hour_by_hour_cpu_need = None
-        self.hour_by_hour_ram_need = None
-        self.hourly_job_occurrences_across_usage_patterns_per_job = ExplainableObjectDict()
-        self.hourly_avg_job_occurrences_across_usage_patterns_per_job = ExplainableObjectDict()
-        self.hourly_data_upload_across_usage_patterns_per_job = ExplainableObjectDict()
         self.server = server
         self.storage = storage
         if not base_ram_consumption.value.check("[]"):
@@ -36,16 +31,11 @@ class Service(ModelingObject):
 
     @property
     def calculated_attributes(self):
-        return [
-            "hour_by_hour_ram_need", "hour_by_hour_cpu_need", "storage_needed"]
+        return ["storage_needed"]
 
     @property
     def modeling_objects_whose_attributes_depend_directly_on_me(self) -> List[ModelingObject]:
         return [self.server, self.storage]
-
-    @property
-    def jobs(self):
-        return self.modeling_obj_containers
 
     @property
     def usage_patterns(self):
@@ -55,20 +45,9 @@ class Service(ModelingObject):
     def systems(self) -> List:
         return list(set(sum([up.systems for up in self.usage_patterns], start=[])))
 
-    def compute_hour_by_hour_resource_need(self, resource):
-        resource_unit = u(self.resources_unit_dict[resource])
-        hour_by_hour_resource_needs = EmptyExplainableObject()
-        for job in self.jobs:
-            hour_by_hour_resource_needs += (
-                job.hourly_avg_occurrences_across_usage_patterns * getattr(job, f"{resource}_needed"))
-
-        return hour_by_hour_resource_needs.to(resource_unit).set_label(f"{self.name} hour by hour {resource} need")
-
-    def update_hour_by_hour_ram_need(self):
-        self.hour_by_hour_ram_need = self.compute_hour_by_hour_resource_need("ram")
-
-    def update_hour_by_hour_cpu_need(self):
-        self.hour_by_hour_cpu_need = self.compute_hour_by_hour_resource_need("cpu")
+    @property
+    def jobs(self):
+        return self.modeling_obj_containers
             
     def update_storage_needed(self):
         storage_needed = EmptyExplainableObject()
