+++ tests/hardware/servers/test_server_base_class.py
@@ -39,6 +39,41 @@ class TestServerBaseClass(TestCase):
         )
         self.server_base.dont_handle_input_updates = True
 
+    def test_update_hour_by_hour_cpu_need(self):
+        job1 = MagicMock()
+        job2 = MagicMock()
+
+        job1.hourly_avg_occurrences_across_usage_patterns = SourceHourlyValues(
+            create_hourly_usage_df_from_list([10, 20, 1, 0]))
+        job2.hourly_avg_occurrences_across_usage_patterns = SourceHourlyValues(
+            create_hourly_usage_df_from_list([20, 15, 5, 3]))
+        job1.cpu_needed = SourceValue(2 * u.core)
+        job2.cpu_needed = SourceValue(3 * u.core)
+
+        with patch.object(Server, "jobs", new_callable=PropertyMock) as mock_jobs:
+            mock_jobs.return_value = {job1, job2}
+            self.server_base.update_hour_by_hour_cpu_need()
+
+        self.assertEqual([80, 85, 17, 9], self.server_base.hour_by_hour_cpu_need.value_as_float_list)
+
+    def test_update_hour_by_hour_ram_need(self):
+        job1 = MagicMock()
+        job2 = MagicMock()
+
+        job1.hourly_avg_occurrences_across_usage_patterns = SourceHourlyValues(
+            create_hourly_usage_df_from_list([10, 20, 1, 0]))
+        job2.hourly_avg_occurrences_across_usage_patterns = SourceHourlyValues(
+            create_hourly_usage_df_from_list([20, 15, 5, 3]))
+        job1.ram_needed = SourceValue(2 * u.GB)
+        job2.ram_needed = SourceValue(3 * u.GB)
+
+        with patch.object(Server, "jobs", new_callable=PropertyMock) as mock_jobs:
+            mock_jobs.return_value = {job1, job2}
+            self.server_base.update_hour_by_hour_ram_need()
+
+        self.assertEqual([80, 85, 17, 9], self.server_base.hour_by_hour_ram_need.value_as_float_list)
+
+
     def test_available_cpu_per_instance_single_service(self):
         service = MagicMock()
         service.base_cpu_consumption = SourceValue(2 * u.core)
@@ -93,8 +128,8 @@ class TestServerBaseClass(TestCase):
         ram_need = SourceHourlyValues(create_hourly_usage_df_from_list([0, 1, 3, 3, 10], pint_unit=u.GB))
         cpu_need = SourceHourlyValues(create_hourly_usage_df_from_list([2, 4, 2, 6, 3], pint_unit=u.core))
 
-        with patch.object(self.server_base, "all_services_ram_needs", new=ram_need), \
-                patch.object(self.server_base, "all_services_cpu_needs", new=cpu_need), \
+        with patch.object(self.server_base, "hour_by_hour_ram_need", new=ram_need), \
+                patch.object(self.server_base, "hour_by_hour_cpu_need", new=cpu_need), \
                 patch.object(self.server_base, "available_ram_per_instance", new=SourceValue(2 * u.GB)), \
                 patch.object(self.server_base, "available_cpu_per_instance", new=SourceValue(4 * u.core)):
             self.server_base.update_raw_nb_of_instances()
@@ -119,8 +154,8 @@ class TestServerBaseClass(TestCase):
         expected_data = [0.5, 1, 1.5, 1.5, 5, 0.5, 1, 1.5, 1.5, 5]
         expected_max_date = start_date_b + timedelta(hours=(len(ram_need_b)-1))
 
-        with patch.object(self.server_base, "all_services_ram_needs", new=all_ram_need), \
-                patch.object(self.server_base, "all_services_cpu_needs", new=all_cpu_need), \
+        with patch.object(self.server_base, "hour_by_hour_ram_need", new=all_ram_need), \
+                patch.object(self.server_base, "hour_by_hour_cpu_need", new=all_cpu_need), \
                 patch.object(self.server_base, "available_ram_per_instance", new=SourceValue(2 * u.GB)), \
                 patch.object(self.server_base, "available_cpu_per_instance", new=SourceValue(4 * u.core)):
             self.server_base.update_raw_nb_of_instances()
